+++ UnitTests/View/Layout/Dim.CombineTests.cs
@@ -1,7 +1,7 @@
 ﻿using Xunit.Abstractions;
 using static Terminal.Gui.Dim;
 
-namespace Terminal.Gui.PosDimTests;
+namespace Terminal.Gui.LayoutTests;
 
 public class DimCombineTests (ITestOutputHelper output)
 {
