+++ UnitTests/View/Layout/Dim.FillTests.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit.Abstractions;
 
-namespace Terminal.Gui.PosDimTests;
+namespace Terminal.Gui.LayoutTests;
 
 public class DimFillTests (ITestOutputHelper output)
 {
