+++ UnitTests/View/Layout/Pos.CenterTests.cs
@@ -3,7 +3,7 @@ using Xunit.Abstractions;
 using static Terminal.Gui.Dim;
 using static Terminal.Gui.Pos;
 
-namespace Terminal.Gui.PosDimTests;
+namespace Terminal.Gui.LayoutTests;
 
 public class PosCenterTests (ITestOutputHelper output)
 {
