+++ UnitTests/View/Layout/Pos.CombineTests.cs
@@ -3,7 +3,7 @@ using Xunit.Abstractions;
 using static Terminal.Gui.Dim;
 using static Terminal.Gui.Pos;
 
-namespace Terminal.Gui.PosDimTests;
+namespace Terminal.Gui.LayoutTests;
 
 public class PosCombineTests (ITestOutputHelper output)
 {
