+++ UnitTests/View/Layout/Pos.FuncTests.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit.Abstractions;
 
-namespace Terminal.Gui.PosDimTests;
+namespace Terminal.Gui.LayoutTests;
 
 public class PosFuncTests (ITestOutputHelper output)
 {
